Fix bug in shrinking array-as-vecs
When shrinking an array-as-vec, the new allocated size was set to almost
exactly the current array size, so adding a few elements would cause it to
grow again, and then removing a few to shrink again.

Instead, wait until the array size is allocatedSize/(growthFactor**2) and then
shrink it down to allocatedSize/growthFactor.