Allow all the tests I just created to pass
Add a new flag, FLAG_SUPER_TEMP, which is attached to temporaries that store
references to the super field of a class.  Then, when deciding if we should
perform a virtual method call, if we know we're in a call on the super field
we can just enter the "dynamic dispatch" code path and force ourselves to enter
the correct branch of the if statement that would be created there.