Update perf playground to test a less "minimal" jemalloc
Currently the perf playground is testing jemalloc configured with:

    --disable-stats --disable-fill --disable-valgrind

but the `fill` and `valgrind` options are required to use jemalloc under
valgrind, so we wouldn't want to outright disable them for production builds.
Just test the performance of

    --disable-stats

to see if it still results in a performance improvement because that's a
feature we don't currently use and could consider disabling.