Deferred actual denormalization
This commit separates the pass into two. Cases where use and defExpr are
CallExprs are deferred to the second pass, which goes through actuals of
CallExprs in reverse order.

There is certain amount of repetition, which increases the pass time a bit. It
is easy to workaround that to shave off some fraction of a second.

Currently, some checks are a bit conservative some are too loose.