First draft of rewriting zippered iteration to avoid tuples
Long ago, Chapel used iteration over a tuple as the means of
expressing zippered iteration.  Then the zip() keyword was
introduced to remove this subtlety and to permit iteration
over tuples themselves.  Implementation-wise, the compiler
still used tuples to represent the set of things to be
iterated over, just to minimize the amount of code that needed
to change (at least, that's how I see/remember it).  This
has caused some problems such as (a) we don't support tuples
that mix types and values, so if you have a type that supports
iteration (as enums would like to), you can't zip them with
things like ranges 1..n; (b) tuples continue to have somewhat
vague semantics in some cases, and as we mess with those
issues, zippering is affected even though it should not be.
It seems likely that there's also potentially some generic
instantiation overhead for using tuples, depending on how
far they get through the compiler.

Here, I'm investigating replacing the tuple with a primitive
which seems more fitting anyway, with the idea that the
primitive is just a placeholder for a list of expressions
until we rewrite the loop.