Add a config param debugArrayAsVec to print resizing messages and add a test
With debugArrayAsVec=true, messages are printed when a push/pop causes the
array's backing memory to be resized.

Added a test that sets debugArrayAsVec=true and does push_back a bunch of
times then does pop_back the same number of times. Verified that both
growing and shrinking the backing array work as expected.