Fix mistake I introduced in resizeAllocRange
When shrinking the allocRange, I mistakenly used a tightly bound range to
compute the new value instead of the previous allocRange. This could cause
it to deallocate memory on the opposite side of the array than the one that
should be changed.

The misuse of the tightly bound range was the only use of it, so removed it
as an argument to the function.