Add check that parentCall is not null
Coverity complained that we shouldn't try to call a method on a null pointer,
which is fair (though I don't think it would necessarily be null, but who knows
what code will be created?).  So insert a check that the parent call is not
null before calling the method on it.