Start at simplifying tuple code
In my arrays branch, I need to adjust the behavior of tuples so that
sometimes they store elements by ref and other times by value. This is
difficult to do in a clear and understandable way with the current
framework for constructing the essential tuple functions (such as type
constructor and _build_tuple). I'm trying to change the tuple
implementation to create these functions in a straightforward way (but
one that re-uses less code).