Create new chplspell-venv Makefile rules, and update documentation
chplspell-venv is based on (copied directly from) the test-venv and
chpldoc-venv.  Except those are user-facing venvs, and chplspell-venv
is developer-only.  So it's set up and described in .devel versions of
the Makefiles and READMEs, where appropriate.

chplspell-requirements.txt drives the venv creation.

Update the LICENSE and third-party/chpl-venv/README.devel.md with
licensing information about the scspell program installed into the
chplspell-venv.

Update gen_release to remove the new developer-only files.