Exclude extern types and possible puts/gets
There is no way of knowing if an extern type is a ptr or not which caused issues
especially with qio* types. I am not entirely convinced that this is the issue
behinf ti but it fixes and it makes sense in the broader sense to prevent
denormalzation of extern type variables.

chpl_gen_comm* functions don't return the value but put the communicated data in
an argument. Therefore a temporary cannot be denromalized if it's def might
trigger a communication.

Also, chpl_gen_comm* functions are generated at codegen time and there is no
direct way to distinguish if a move with primitive on rhs might trigger a
communcation. This commit implements a function to check if a move can trigger
comm.

This commit also does a cleanup.