Refactor codegenPrimMoveRhsIsSpecialPrimop
This function was used as a codegen helper for RHS of a move for special
primitives. Now that the AST is denormalized at codegen time, those special
primitives can appear anywhere. This commit modifies the said function so that
it can be run from different places.

Currently callsites are located in codegenPrimMove(as before) codegenPrimitive,
where some of those primitives were listed in a case fall-through, and
CallExpr::codegen where an actual can be one of these primitives.

Interim commit for helper refactoring

Finish refactor