Replace some checks against the name of the "meme" arg with flag checks
Checking a flag that is already there is preferable to performing string
comparisons.

This patch assumes that all "meme" arguments will have the flag FLAG_IS_MEME
attached to them.  I performed a paratest over the std build with asserts to
verify this assumption and saw no failures, so I believe it is a fair
assumption.  I also observed no code paths which added the argument without the
flag.