Condense similar branches in scopeResolve check
I realized that the behavior within the if statement branches was identical and
that the only difference between the checks was the order - we're verifying
that a function doesn't conflict with a named symbol of another type (like a
variable, or a module, etc) and the difference between the two checks was
whether the function was the new symbol, or the old one.  Since they accomplish
the same thing, make them one check in the if statement and drop the else if
branch.