Add perf tracking for all kernels, variants and loop sizes currently in LCALS
Instead of running parts of LCALS independently, enable all of it to run in
a single go. Set the new config param perfTesting=true to output times in an
easier to parse format.

Data for all of the perfkeys will now end up in a single .dat file, with
multiple graphs generated from the data it contains.

Added a longer timeout for LCALS since it takes a while to run the whole
thing.  We can reduce the timeout value once we know how long it should
take in all perf configurations.

Removed the .prediff that was used to massage the output format into
something easier to parse.