Add annotation to performance graphs for incremental comp codegen changes
Some of the changes made in preparation for --incremental resulted in slower
code during performance testing.  This is due to no longer applying the static
keyword to functions within the chpl__header.h file.  We are planning on
reapplying that keyword during normal compilation, but will not restore it for
incremental compilation because otherwise it will not work at all.