improve operator overloads, apply multi locale bug fix, general cleanup
Adds operator overloads with lhs int arguments for most operators. op/
and op% still do not take lhs int arguments.

Multi-locale bug fix is applied to all functions, instead of just the
writeThis  and get_str functions.

Cleans up test formatting, moves div_ui out of the record since it
doesn’t modify the “this” BigInt it was being called on. Removes unused
maybeCopy and localize functions. Renamed the seed(uint) function to
seed_ui to match the naming conventions for the rest of the module.
Remove the “simple” test now that the bug hunt it was made for is done,
since it doesn’t cover any module code not covered in other tests.