Update removeUnnecessaryAutoCopyCalls.cpp for ref changes
removeUnnecessaryAutoCopyCalls.cpp previously assumed that the argument
to autoCopy was the same type as the result.  That's not true in the
event that autoCopy is taking an argument that is a ref (it can do that
b/c blank intent is const ref for records e.g.). So, we replace with a
PRIM_DEREF in that case. Since we're removing an autoCopy, adding the
PRIM_DEREF here does not seem to come at a risk to correct record
behavior.