Attempt to reconcile all the LD settings
Too many Makefiles try to set LD in an uncoordinated manner.
The ones that win _seem_ to be the ones in Makefile.include and
Makefile.launcher, respectively (where the former should probably
be moved to Makefile.exe for symmetry; or Makefile.launcher should
include Makefile.include).

However, GASNet would like to specify the linker, and in a recent
case, a user had problems because he wasn't getting mpicc as the
linker as GASNet requested.  So maybe all the comm layers should
get to choose?

However, regexp-re2 also seems to want to set it.  Michael was
concerned that if the C++ code in re2 wasn't linked with CXX,
we'd end up missing the libstdc++ libraries.  However, that doesn't
seem to be the case in practice (i.e., it looks to me like
the Makefile.include setting of simply CC wins, without issue).

Strangely, if I let GASNet win, on my Mac, it chooses clang++
which generates a bunch of warnings of the form:

ld: warning: object file
(/path/to/libchpllaunch.a(arg.o)) was built for newer OSX version (10.11) than
being linked (10.10)

I'm timing out and moving on for now.