Remove clone_parameterized_primitive_methods and handle it in canInstantiate
David pointed out that a line I'd dismissed in function resolution actually
accomplished the same purpose for dtIntegral, so the entire clone_parameterized
function could be removed, with the complex handling getting replaced by a
similar two line check in canInstantiate.  This works for a couple of spot
checks, and full testing will get run momentarily.