Update how remoteValueForwarding finds functions with sync var operations
remoteValueForwarding previously found sync var operations by looking for the
sync var primitives and then finding functions that called the primitives.
However, we're working towards removing the primitives and just calling extern
functions directly.

Instead of looking for functions with sync var primitives in them, we now just
look for functions whose "this" argument is a sync var (i.e. we're finding
methods on sync vars)

Looking at the compiler output for release/examples/primers/syncsingle when
DEBUG_SYNC_ACCESS_FUNCTION_SET is defined shows that we're finding the same set
of functions that we were before this patch. In addition, we're also finding
the destructors on sync vars now, but I don't think that's a bid deal. In other
words, this just updates how we're finding functions with sync var operations,
it's not really changing the end behavior.