Set dataParMinGranularity for LCALS
The parallel variant of LCALS was using way too many tasks for the Medium and
Small loop sizes in LCALS on systems with larger numbers of cores, making the
tasking overhead swamp the actual computation, causing perf testing timeouts.

By setting the dataParMinGranularity to 1000, a system with 24 cores that was
going past 3000 seconds to time out is now under 400 seconds.

While there, also set -sassertNoSlicing to improve the serial variant.