Add a version of BinaryTrees that uses ** instead of bitshift
Adds the chapel, good, perfexecopts and perfkeys files for the new
study version (based on the release version).

New study version is to compare the performance of 2 ** (expression)
with the current 1 << (expression) implementation.

The little bit of testing that I did didn’t show a performance
difference greater that couldn’t just be noise.