Stop printing "Error" when tags utility is not found during make in modules
If CHPL_DEVELOPER is on, then if there isn't a tags utility that our build
system knows how to use, then the last step of building from $CHPL_HOME:
"cd modules && make" prints "make[2]: [all] Error 1 (ignored)" or similar
as the last line of output from a successful build.  This is confusing.

Add " || true" to the command that causes the error, so it is silently
ignored instead of confusingly ignored.