Quiet a remote value forwarding future
Quiet a remote value forwarding future by making sure the on statement starts
executing before we update the scalar that is being mistakenly forwarded. The
forwarded value is put into the arg bundle for the on-stmt so we need to make
sure that the value isn't updated before the arg bundle is created, otherwise
we're forwarding the updated value, which makes it appear as though the test is
passing.