Improved #line directives in generated C code.
CHANGES IN GENERATED CODE

This PR results in the following changes in the generated C code,
when applicable due to --cpp-lines and --gen-ids compiler options:

* All "id comments", i.e. /* NNN */ where NNN is an AST node's ID,
are generated uniformly by idCommentTemp() using %7d format.
idCommentTemp() returns a static buffer for simplicity, hence "Temp".

* In chpl__header.h, each field of a struct type has an id comment.
I preferred not to generate a #line per field for brevity
and to avoid inadequate #line diagnostics for adjacent things.

* In chpl__defn.c, each global variable declaration corresponding
to a user variable has an id comment and a #line.

* In per-module .c files, the #line directive for each function and local
variable declaration reflect the astloc from the corresponding ast node.

* There is a #line 0 "<internal>" directive at the top of _main.c,
to reduce user exposure to the generated code upon --cpp-lines.

* There is also a #line 0 "<internal>" directive after the declarations
of global user variables in chpl__defn.c. Avoids pointing, for internal
things like chpl_mem_descs, into whatever file the last user variable
was in.

As an aside, my preference is to improve things further so chpl__header.h:

* contains an id comment and a #line for each struct type,

* does not contain an id comment for things that are defined in .c files,
especially function prototypes;

* separately, force no-#line mode when CHPL_CG_CPP_LINES=n is set
and no #line-related options are given on the command line,
even when CHPL_DEBUG=y.

I suggest - and include a related change to beautify.cpp - that we do not
necessarily emit #line 0 "<internal>" for consecutive lines.
The benefit is slightly smaller generated code. For example,
chpl__defn.c contains #lines only for global-user-variable decls.
The cost is that the C compiler may report "<internal>:nn" for nn>0,
besides just "<internal>:0".

Minor - we get a slight duplication right before a function header:

#line 22 "DefaultRectangular.chpl"
/* DefaultRectangular.chpl:22 */
#line 22 "DefaultRectangular.chpl"
void chpl__init_DefaultRectangular(int64_t _ln, int32_t _fn) {

DETAILS

#line directives are added by beautify(fileinfo* origfile), replacing
the comments /* ZLINE: lineno filename */ generated by codegen.

I think that when beautify() sees a ZLINE line, it records (lineno, filename).
When it sees any other line, it outputs a #line with the currently-recorded
(lineno, filename), if any, prior to printing the line itself.

Prior to this change, ZLINE comments were added before each statement
within a function and were not added before the function header.
Therefore the first function in a generated .c file had no #line
directives - until the first statement in its body. Each subsequent
function started out with #line directives that duplicated the last
#line in the preceding function.

Now we are generating a ZLINE before the header of each function
and before each local variable declaration. That way their #line
directives are meaningful, not just cloned from the previous line.

There are still some ZLINE comments left unhandled by beautify().
This was the case prior to this change as well.
This is independent of --gen-ids.

These unhandled comments can be found at least at "then" and "else"
branches of conditional statements. I think this happens whenever
we start codegen-ing a BlockStmt not at start of the line
of the generated C code. This is because beautify() recognizes ZLINE
comments only at the beginning of the line - see

  if (!strncmp(cp, ZLINEINPUT, ZLINEINPUTLEN))

Probably related to those cases where we generate to node-ID comments
at the beginning of a line.

I did not attempt to fix these cases.

BTW "ZLINE" is a cute historical artifact. I think it's short for "ZPL line"
i.e. line number in the code written in ZPL. This exposes code lineage,
see the comment near the top of beautify.cpp.

This is a replacement for #4260 to resolve merge conflicts.