Remove chpl_task_yield from meteor-parallel-alt's atomic lock/unlock
This is basically a duplicate of #4226. #4226 had to be backed out because of a
deadlock under our memory tracking testing. This skips meteor-parallel-alt for
memory tracking and adds a futre to catpure the problem. The basic issue is
that the atomic lock/unlock requires that tasks aren't yielding Memory tracking
currently uses a sync var as a lock, and sync vars can occasionally yield
(depending on the tasking layer), which causes deadlock.