Protect against divide by zeroes
In a recent Chapel training, we had users accidentally send in all-zero
arrays to the MPlot routines, which resulted in a fault due to dividing
by zero.  This mod seeks to circumvent that by checking that the number
we're dividing by is non-zero.

Given the unfortunate surprise of getting a segfault in Chapel, we
might consider putting such a check in the divide implementations
themselves when checks are off, but I'm not undertaking that here...