Fix some code that was broken and probably never executed.
In my initializers work, I went down into this else branch in function
resolution's insertCasts() and discovered that this branch did not properly
execute on its own.  The PRIM_MOVE inserted tried to use a AST node which was
already in the AST (as was made clear by the ->replace() call which followed
the else branch) and that is an error in our compiler.  After fixing that,
there was a complaint about encountering a use before resolving the def, because
the DefExpr inserted in that else branch never had resolve called on it (nor
did any of the other Exprs inserted in this entire branch).  So I inserted some
resolveExpr calls, too.

I don't think any of this is executed, but since I'm running into it now, it
seems reasonable to try to maintain it.  I'm still encountering bugs in my code
- if more of them seem to be related to this else branch being stale, I will
make further fixes.