revcomp cleanups related to chunk processing
This commit cleans up the logic around finding the section starts
and processing the sections in the input data of revcomp.
Specifically, it reworks some of the logic around finding the
start of a chunk and processing it, which permits us to remove
the logic to skip past the header information by simply skipping
it in the indices that we send into process().

It also guards the final call to process() in the event that the
input file is empty or we never found a chunk to process.

On our XC testing, consoleIn.length() returns 0, which makes this
program generate an out-of-bounds error trying to access the (empty)
data[] array.  This commit does not fix that issue, but cleans up
some of the logic around finding the section starts