Workaround for problems with new C(A=>GA)
This pattern is used in several distribution implementations.
See test_array_alias_field.chpl for a short test case.

It's not clear to me that we want to continue to support this
feature but this relatively small change seems to enable it
to continue to function with the new array implementation.