Yield less frequently in meteor's testAndSet lock
In meteor-parallel-alt we use a testAndSet lock to serialize/protect calls to
recordSolution(). Normally we'd just use a syncvar as a lock, but that hurts
performance for this insanely fast benchmark. The lock has minimal contention
and sync vars are best for highly contended locks. In this case using as sync
var almost doubles the execution time.

Previously a testAndSet loop with a yield every failed lock attempt was used.
Unfortunately this isn't ideal either, because yielding will lead to
unnecessary work and cache misses. Ideally we'd just do a pure testAndSet loop
without yielding, but that's not possible. Under qthreads (a non preemptive
scheduler) if some other thing (like memory tracking) yields the qthread, then
a plain testAndSet lock will deadlock, so we need to yield at some point to
avoid starvation.

For meteor-parallel-alt on the shootout machine, we need to spin ~2**16 times
before yielding to avoid yielding under perf runs. Unfortunately, that makes
meteor take ~300 seconds with memory tracking turned on. Basically if memory
tracking yields the qthread, we have to wait for all ~1,000 tasks to spin wait
2**16 times to reschedule the task that needs to unlock the lock.

To work around these problems, this patch uses an exponential backoff lock that
decreases the number of lockAttempts until the next yield. This gives us the
best performance (.065 seconds compared to a previous .08 seconds) while still
allowing memory tracking to finish quickly (.2 seconds.) This should put us in
the top 5 for the shootout entries.