Ensure formals' types are retained.
In the following code:

  class MyType {}
  proc test(arg:MyType) {}
  test(nil);

where MyType is otherwise unused, e.g. no instances or subclasses of
it are created, MyType used to get pruned during resolution by
removeUnusedTypes() because isUnusedClass() would return true for it.

This changes adds a sweep of formals to ensure that isUnusedClass()
returns false for MyType in the above example. To do so, I ensure
that ct->defaultTypeConstructor is resolved.

I also considered these alternatives:

* Add another flag like "this type is used" and have isUnusedClass()
check for this flag. ==> I was happier to find a solution that
does not require a new flag.

* Mark MyType with FLAG_GLOBAL_TYPE_SYMBOL. ==> This flag
seems to apply only to predefined types, I did not want to mess
with that.

By contrast, resolving defaultTypeConstructor seems benign
and makes sense as in "the type is used so the type constructor
may as well be usable".

I placed the sweep of formals, aka ensureResolvedArgTypeConstructors(),
earlier that removeUnusedTypes - at start of removeUnusedFunctions().
Otherwise the latter observes that the defaultTypeConstructor is
unresolved and prunes it. We could re-insert it into the AST if that
happens, however we'd need to know where, and that gets lost when the
function is removed from AST.

Doing the sweep before functions are pruned may result in
us tending to a formal of a function that will end up pruned.
I think this is OK and in practice will not be a big deal.

In writing ensureResolvedArgTypeConstructors(), I considered
the following issues:

* Its placement - see above.

* Which functions' arguments to ignore - I added the checks
until it worked. The functions must be non-generic and
have been resolved. In retrospect, the test for non-genericity
is not needed: generic functions are never resolved.

* What types not to deal with. I ignored most types that
isUnusedClass() already returns false for. The types
I would still consider are ones that have subclasses.
This should be not a big deal.

* Given that most formals will not need the adjustment,
what is the order of checks that will allow us to detect
that the fastest - to reduce the amount of time we spend
in the sweep.

The bug was reported by Vitali Baumtrok.

Passes full suite under linux64 and SSCA2 under gasnet.