In build.cpp, move iterableExpr to call to parloopexpr()
This provides a solution to a problem I ran into while working on array
improvements. With my array improvements, iterator records capture array
and domain arguments by ref (since autoCopy now does a deep copy and
iterator records are copied a number of times in normal use). That change
causes problems for cases like the below:

    const a3 : [1..6,1..3] int = [(i,j) in {1..6,1..3}] i+j;

    (from test/users/shetag/unboundslice2.chpl ).

in which the expression {1..6,1..3} is a domain that is created in a
"parloopexpr" function (created in build.cpp).  But the "parloopexpr"
function returns an iterator record, which the caller iterates over. The
problem is that - if iterator records capture arrays and domain arguments
by ref - by the time "parloopexpr" returns, the anonymous domain has
already been freed (and the iterator record now refers to garbage
memory).

This change adjusts build.cpp to generate the iterable expression such as
{1..6,1..3} in this case in the call site for parloopexpr and for the
body of parloopexpr to use it as an argument.