Add enum tag to Aggregate Types to differentiate based on constructor style
Defines three cases for the enum:
- Haven't seen a constructor or initializer yet, use compiler defined one
- Seen an old style constructor
- Seen the new initializer form

Sets the initial state of an AggregateType to "haven't seen either yet".  This
will allow us to verify a user hasn't defined both an old style constructor and
the new style initializer for their class, by having the first one set always
impact this field so that any following definitions will check against it.