Enable, refactor, and improve the strided bulk transfer optimization.
Sets `useBulkTransferStride` to `true` by default.

Significantly refactors the existing implementation for the strided bulk
transfer optimization. A number of utility and helper functions seemed
overly complex, and I felt that I could express them in a simpler way.
This new refactoring relies on the observation that we only need a new
stride level if the current stride value cannot be re-used to advance to
the next chunk of elements. With this in mind, we really only need the
helper function now named `canReuseStride` in order to compute the
count, stride values, and stride levels.

I consider this refactoring to be much more intuitive, and seems to have
fixed a few bugs in the previous implementation.

In an attempt to reduce communication, we now move to the locale of the
LHS of the assignment if neither array lives on the current locale. This
also simplifies the function that invokes the runtime calls that
actually do the transfer.

Minor changes:
- Updates the reference paper to a cleaner-looking URL.
- Cleans up some whitespace floating around.