Add some tests of the new error message
I realized I didn't have a test case with both an old style initializer and a
new constructor.  Prior to this change, if the type designer defined one of each
with the same argument list, we'd have an ambiguous call when the type was
used, but if they had different arguments we wouldn't mind.  With the change,
these tests now print out the same error message.