Modify assertion in captureTaskIntentValues()
captureTaskIntentValues() contained assertions
one of which was failing under gasnet with --verify.
This change replaces them with another assertion
which passes under linux64 and gasnet.

The previous assertions examined the structure of the AST,
verifying that varActual is a loop induction variable.
I replaced that with a direct check of FLAG_COFORALL_INDEX_VAR.
Verifying the particular structure of the AST does not seem useful
in this case.

The reason the failure had gone unnoticed for a while, perhaps since
that code was added, is because the code is run only upon --verify,
and we do not require that compiler executes successfully with
--verify under gasnet.