Change the internal name of the init definitions to "init" and call them
In buildDefaultFunctions, updated record defaultOf function to call "init"
instead of "_construct_<typename>" when there is an "init" function defined on
the type, as that should be the name of the default function.

In normalize, stop changing the name of the init function, but still change the
name of constructors.

In functionResolution's resolveNew, call "init" if there are initializers
defined on the type instead of "_construct_<typename".

In misc.cpp's printErrorHeader, continue to output "In initializer" for "init"
functions, which would otherwise be labelled as just "functions" in the output.
Update the expected output for constAndParamInHeader.chpl as a result of this.

This commit has two failures:
- classes/initializers/generic-uninit-field-initializers.chpl
- classes/initializers/record-var-invokes-user-initializer.chpl
both of which are a result of being unable to differentiate between multiple
init functions (because their argument lists look identical).  This will be
addressed next.