Added a small cleanup that Paul noticed
The final loop of the program is more cleanly written as a strided
loop.  I think it looked this way because I had made it a param
loop in my study version, yet we can't do param loops over strided
ranges :'(, so I put the conditional in the loop to work around it.
Paul's approach is cleaner.