Fix a bug in flattenNestedFunctions
flattenNestedFunctions contains logic to support outer functions that call
inner functions (such as toFollower/toLeader in ChapelIteratorSupport).
However, this logic contained an error.  The case that worked incorrectly was:

 * inner function chpl__reduce4 calls outer function toFollowerZip
 * outer function toFollowerZip calls inner function _iterator_for_loopexpr3
 * inner function _iterator_for_loopexpr3 uses an outer variable k

In that case, _iterator_for_loopexpr3 and toFollowerZip would get an argument
for the outer variable k, but chpl__reduce4 would not. That would lead to
errors in C compilation in which a k in chpl__reduce4 refered to a variable not
in that function.

To fix the error, I simply moved the logic identifying outer functions with
captured variables inside the first do { } while(change) loop since that loop
already handles the case of adding captured variables for nested functions.

(Note that previous parts of the compiler probably do not yet generate this
pattern, but they will after the next commit).