In build.cpp, pass dataExpr as argument to reduce/scan functions
This is a follow-up to PR #4310. For the same reasons as in that PR (ongoing
work with arrays), this PR adjusts chpl__scan and chpl__reduce functions to
take in the reduced/scanned expression by reference. That way, when the
reduced/scanned expression allocates an object, it will be destroyed only after
iteration over it has completed (when the iterator returned by chpl__scan is
run).

The test reductions/bradc/minmaxlocscan.chpl is an example that needs this
adjustment with the new array implementation.