Change initializers to take in the memory to initialize.
Moves alloc calls to the site of the "new" call on types that define our new
initializers.  Insert the meme argument into calls from record _defaultOf and
initCopy functions, as well as the compiler-inserted user initializer calls to
the parent initializer and calls from the type constructor.  In
buildDefaultWrapper, skip the alloc call insertion if we define initializers,
avoid dereferences of NULL when handling the meme argument, and pass the meme
argument on straight.