Add error message when new initializers used with generics
We'd like this to work eventually, but unfortunately making the initializer
take in the memory to initialize made it obvious that we didn't know the
generic instantiation before the call was resolved.  I think ideally we'd like
to change what happens with type constructors so that we'd resolve them before
the constructor call rather than while resolving the constructor call, but that
will require a lot more code changes and anyways we haven't finalized our
generic story yet.

Make the generic test that was "passing" a future again.