Add an error if a constructor or initializer is declared without parentheses
A future from 2011 indicated that at least Brad and Vass thought this error
was a reasonable idea.  The fix was really simple - when we're converting a
method into a constructor, just check if it was declared with parentheses.

Had to update the line number in the initializer version of the future.  Removed
the .bad and .future files.

Conflicts:
	compiler/passes/normalize.cpp