Add an optimization to ** for param bases that are powers of 2
Adds a version of the ** operator for the case where the LHS (base) is
known at compile time and a positive power of two. Uses bit-shifting to
efficiently raise the power.

Uses complement and compare as an efficient method of testing whether
the param is a power of 2.

Refactored the body of the overload into _expBaseHelp to match the
style of the nearby functions. Naming help welcome.

I have a couple of comments about how things work so that people who
don’t understand bits well can follow along. Should internal module
code have comments beyond general section comments?