Fix an issue when requesting a new capacity for an associative domain
The following test would fail when its random number generator happened
to request a capacity of '7':

domains/johnk/capacityAssoc.chpl

The error message would complain about an out-of-bounds access on an
array with the index '-1'. I tracked this value down to one of the
returns in the "_add" function. The simple fix is to return the found
slot instead of a sentinel '-1' value.