Necessary normalize change to .init calls
Constructors and initializers are different from other methods - they don't take
in the _mt token.  Normally we don't have calls to <instance>.<constructor>(),
so normalize's applyGetterTransform doesn't cause problems, but with the new
initializer story we rely on the presence of this.init() and super.init() calls.
Additionally, initializers now take in their memory to initialize as the final
"meme" argument (to help distinguish between inits of records which have similar
field set up), and ordinary method calls have the instance of the type to
operate on in the first argument.  Thus two adjustments need to occur (and could
occur at the same time): 1) .init() calls should not be given the _mt actual
and 2) the first actual should be made into a named actual.

With this minor modification, three of my futures now pass, and many more of
them are now only encountering problems with super.init() calls when super ==
object.  Retire the .future files on those tests.