Unset CONFIG_SITE, so as to have dependable third-party install paths.
CONFIG_SITE is an environment variable used to specify site-specific
autoconf defaults.  Unfortunately, use of it can interfere with our
assumptions about third-party builds and installs, including where we
think the installed libraries are placed.  To work around this, for now
prevent it being passed along to the environment of any subprocess of
our make.

See https://sourceforge.net/p/chapel/mailman/message/33852491/ for a
conversation about this.