Basic migration to QualifiedType
 * replaces the virtually-dispatched typeInfo() with a function,
   called qualType(), that returns a QualifiedType.
 * implements typeInfo() using qualType() in baseAST
   (this way, I didn't need to change the uses of typeInfo() yet)
 * leaves Symbol->type alone (to avoid updating uses of it yet)
   but adds a Qualifier sibling to it
 * Adjusts Symbol::qualType() to translate argument intents
   into Qualifier.

I did my best to approximate what I expect to be the solution,
but this version is transitional. Notably, nothing actually uses
the Qualifier field in the QualifiedType yet, and so this change
is more about setting up the right structure. There should be
no change in behavior.

One might be tempted to make FnSymbol::qualType() return the function's
return type. However, that's probably not right when we support
first-class-functions or functions passed as arguments. So, there
is a separate method FnSymbol->getReturnQualType() to get the return
type which is called by CallExpr->qualType().

Additionally, depending on how further development with this change
goes, we might consider changing the way that reference types are
handled. For now, primitives that return a reference return
QualifiedType(someType->refType, kRef).

In the future, we might wish do either or both of:
 * change the above pattern to QualifiedType(someType, kRef)
 * translate from reference types to QualifiedTypes with qual=kRef.